+++ tests/mainloop/test_mainloop.py
@@ -1,55 +1,6 @@
 import pytest
 from masr.mainloop import mainloop
-import asyncio
 import uuid
-
-
-async def send(url):
-    await asyncio.sleep(1)
-    return "SUCCESS"
-
-
-class Env:
-    def __init__(self):
-        self.url = "http://mas.url"
-        self._success = False
-
-    def init(self, task_config):
-        pass
-
-    async def run(self):
-        return not self._success
-
-    async def send(self, url):
-        await asyncio.sleep(1)
-        return "SUCCESS"
-
-    async def check(self, promise):
-        await asyncio.sleep(1)
-        return Report(status="SUCCESS")  # Mocked success for simplicity
-
-    def set_success_flag(self, tid):
-        self._success = True
-
-
-class Mas:
-    def __init__(self):
-        self.url = "http://env.url"
-
-    def init(self, task_config):
-        pass
-
-    async def send(self, url):
-        await asyncio.sleep(1)
-        return "SUCCESS"
-
-    async def handel_failure(self, report):  # start again if fail
-        print("Handeling failure", report)
-
-
-class Report:
-    def __init__(self, status):
-        self.status = status
 
 
 @pytest.fixture
@@ -59,16 +10,6 @@ def mock_task_config():
     }
 
 
-@pytest.fixture
-def mock_env():
-    return Env()
-
-
-@pytest.fixture
-def mock_mas():
-    return Mas()
-
-
 @pytest.mark.asyncio
 async def test_mainloop_success(mock_task_config, mock_env, mock_mas):
     # Mock methods to always return success
@@ -76,7 +17,7 @@ async def test_mainloop_success(mock_task_config, mock_env, mock_mas):
         return "SUCCESS"
 
     async def mock_check_success(promise):
-        return Report(status="SUCCESS")
+        return "SUCCESS"
 
     mock_env.send = mock_send_success
     mock_mas.send = mock_send_success
@@ -93,7 +34,7 @@ async def test_mainloop_failure(mock_task_config, mock_env, mock_mas):
         return "SUCCESS"
 
     async def mock_check_failure(promise):
-        return Report(status="FAILURE")
+        return "SUCCESS"
 
     mock_env.send = mock_send_success
     mock_mas.send = mock_send_success
